runtime: splitting classes into two files (#9576)
Risk Level: n/a (just moving code)
Testing: n/a
Docs Changes: n/a
Release Notes: n/a
